requirements.txt
faker
python-kafka
